backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.25.2
openalgo==1.0.15
apscheduler==3.10.4
pytz==2023.3
requests==2.31.0
aiofiles==0.23.2